Introduced ConfigController and moved logic from AuthController
