--- app/Http/Controllers/ExampleController.php
@@ -1,18 +1,0 @@
-<?php
-
-namespace App\Http\Controllers;
-
-class ExampleController extends Controller
-{
-    /**
-     * Create a new controller instance.
-     *
-     * @return void
-     */
-    public function __construct()
-    {
-        //
-    }
-
-    //
-}
